Clean up deleteNodes.go

+++ services/builder/deleteNodes.go
@@ -16,39 +16,25 @@ type etcdPodInfo struct {
 }
 
 var (
-	exportCmd = `export ETCDCTL_API=3 && 
+	exportEtcdEnvsCmd = `export ETCDCTL_API=3 && 
 		export ETCDCTL_CACERT=/etc/kubernetes/pki/etcd/ca.crt && 
 		export ETCDCTL_CERT=/etc/kubernetes/pki/etcd/healthcheck-client.crt && 
 		export ETCDCTL_KEY=/etc/kubernetes/pki/etcd/healthcheck-client.key`
 	getEtcdPodsCmd = "get pods -n kube-system --no-headers -o custom-columns=\":metadata.name\" | grep etcd"
 )
 
+//deleteNodes function finds particular nodes for deletion and deletes them from the etcd and k8s clusters
+//function also changes config.Current state after the nodes are deleted
+//return config with new current state and nil if successful, nil and error  otherwise
 func deleteNodes(config *pb.Config, toDelete map[string]*nodesToDelete) (*pb.Config, error) {
 	for _, cluster := range config.CurrentState.Clusters {
-		var nodesToDelete []string
-		var etcdToDelete []string
-		del := toDelete[cluster.ClusterInfo.Name]
-		for _, nodepool := range cluster.ClusterInfo.NodePools {
-			for i := len(nodepool.Nodes) - 1; i >= 0; i-- {
-				val, ok := del.nodes[nodepool.Name]
-				if val.Count > 0 && ok {
-					if nodepool.Nodes[i].NodeType > pb.NodeType_worker {
-						val.Count--
-						nodesToDelete = append(nodesToDelete, nodepool.Nodes[i].GetName())
-						etcdToDelete = append(etcdToDelete, nodepool.Nodes[i].GetName())
-						log.Info().Msgf("Choosing Master node %s, with public IP %s, private IP %s for deletion\n", nodepool.Nodes[i].GetName(), nodepool.Nodes[i].GetPublic(), nodepool.Nodes[i].GetPrivate())
-						continue
-					}
-					if nodepool.Nodes[i].NodeType == pb.NodeType_worker {
-						val.Count--
-						nodesToDelete = append(nodesToDelete, nodepool.Nodes[i].GetName())
-						log.Info().Msgf("Choosing Worker node %s, with public IP %s, private IP %s for deletion\n", nodepool.Nodes[i].GetName(), nodepool.Nodes[i].GetPublic(), nodepool.Nodes[i].GetPrivate())
-						continue
-					}
-				}
-			}
+		//get nodes to delete for this cluster
+		clusterNodesToDelete := toDelete[cluster.ClusterInfo.Name]
+		if clusterNodesToDelete == nil {
+			return nil, fmt.Errorf("cluster %s does not have any nodes to delete", cluster.ClusterInfo.Name)
 		}
-
+		//get node names to delete + possible etcd members to delete
+		nodesToDelete, etcdToDelete := getNodesToDelete(clusterNodesToDelete, cluster.ClusterInfo.NodePools)
 		// Delete nodes from an etcd
 		if len(etcdToDelete) > 0 {
 			err := deleteFromEtcd(cluster, etcdToDelete)
@@ -61,7 +47,6 @@ func deleteNodes(config *pb.Config, toDelete map[string]*nodesToDelete) (*pb.Con
 		if err != nil {
 			return nil, err
 		}
-
 		// Delete nodes from a current state Ips map
 		for _, nodeName := range nodesToDelete {
 			for _, nodepool := range cluster.ClusterInfo.NodePools {
@@ -78,6 +63,7 @@ func deleteNodes(config *pb.Config, toDelete map[string]*nodesToDelete) (*pb.Con
 }
 
 // deleteNodesByName checks if there is any difference in nodes between a desired state cluster and a running cluster
+//return nil if successful, error otherwise
 func deleteNodesByName(cluster *pb.K8Scluster, nodesToDelete []string) error {
 	// get real node names
 	realNodeNames, err := kcGetNodeNames(cluster.Kubeconfig)
@@ -106,6 +92,8 @@ func deleteNodesByName(cluster *pb.K8Scluster, nodesToDelete []string) error {
 	return nil
 }
 
+//deleteFromEtcd function deletes members of the etcd cluster. This needs to be done in order to prevent any data corruption in etcd
+//return nil if successful, error otherwise
 func deleteFromEtcd(cluster *pb.K8Scluster, mastersToDelete []string) error {
 	mainMasterNode := getMainMaster(cluster)
 	if mainMasterNode == nil {
@@ -133,7 +121,7 @@ func deleteFromEtcd(cluster *pb.K8Scluster, mastersToDelete []string) error {
 		for _, etcdPodInfo := range etcdPodInfos {
 			if nodeName == etcdPodInfo.nodeName {
 				log.Info().Msgf("Removing node %s, with etcd member hash %s \n", etcdPodInfo.nodeName, etcdPodInfo.memberHash)
-				cmd := fmt.Sprintf("%s \" %s && etcdctl member remove %s \"", kcExecEtcdCmd, exportCmd, etcdPodInfo.memberHash)
+				cmd := fmt.Sprintf("%s \" %s && etcdctl member remove %s \"", kcExecEtcdCmd, exportEtcdEnvsCmd, etcdPodInfo.memberHash)
 				err := exec.Command("bash", "-c", cmd).Run()
 				if err != nil {
 					log.Error().Msgf("Error while etcdctl member remove: %v", err)
@@ -145,6 +133,8 @@ func deleteFromEtcd(cluster *pb.K8Scluster, mastersToDelete []string) error {
 	return nil
 }
 
+//getEtcdPods finds all etcd pods in cluster
+//returns slice of pod names and nil if successful, nil and error otherwise
 func getEtcdPods(master *pb.Node, cluster *pb.K8Scluster) ([]string, error) {
 	// get etcd pods name
 	podsQueryCmd := fmt.Sprintf("kubectl --kubeconfig <(echo \"%s\") %s-%s", cluster.GetKubeconfig(), getEtcdPodsCmd, master.Name)
@@ -156,19 +146,24 @@ func getEtcdPods(master *pb.Node, cluster *pb.K8Scluster) ([]string, error) {
 	return strings.Split(string(output), "\n"), nil
 }
 
+//getEtcdMembers will find all etcd members in etcd cluster
+//returns slice of etcd member infos and nil if successful, nil and error otherwise
+//
+// Example output:
+// [
+// "3ea84f69be8336f3, started, test2-cluster-name1-hetzner-control-2, https://192.168.2.2:2380, https://192.168.2.2:2379, false",
+// "56c921bc723229ec, started, test2-cluster-name1-hetzner-control-1, https://192.168.2.1:2380, https://192.168.2.1:2379, false"
+// ]
 func getEtcdMembers(cluster *pb.K8Scluster, kcExecEtcdCmd string) ([]string, error) {
 	// Execute into the working etcd container and setup client TLS authentication in order to be able to communicate
 	// with etcd and get output of all etcd members
-	cmd := fmt.Sprintf("%s \" %s && etcdctl member list \"", kcExecEtcdCmd, exportCmd)
+	cmd := fmt.Sprintf("%s \" %s && etcdctl member list \"", kcExecEtcdCmd, exportEtcdEnvsCmd)
 	output, err := exec.Command("bash", "-c", cmd).CombinedOutput()
 	if err != nil {
 		log.Error().Msgf("Error while executing command %s in a working etcd container: %v", cmd, err)
 		log.Error().Msgf("prepCmd was %s", kcExecEtcdCmd)
 		return nil, err
 	}
-	// Example output:
-	// 3ea84f69be8336f3, started, test2-cluster-name1-hetzner-control-2, https://192.168.2.2:2380, https://192.168.2.2:2379, false
-	// 56c921bc723229ec, started, test2-cluster-name1-hetzner-control-1, https://192.168.2.1:2380, https://192.168.2.1:2379, false
 	// Convert output into []string, each line of output is a separate string
 	etcdStrings := strings.Split(string(output), "\n")
 	//delete last entry - empty \n
@@ -178,6 +173,8 @@ func getEtcdMembers(cluster *pb.K8Scluster, kcExecEtcdCmd string) ([]string, err
 	return etcdStrings, nil
 }
 
+//getEtcdPodInfo tokenizes an etcdMemberInfo and data containing node name and etcd member hash for all etcd members
+//return slice of etcdPodInfo containing node name and etcd member hash for all etcd members
 func getEtcdPodInfo(etcdMembersString []string) []etcdPodInfo {
 	var etcdPodInfos []etcdPodInfo
 	for _, etcdString := range etcdMembersString {
@@ -190,6 +187,8 @@ func getEtcdPodInfo(etcdMembersString []string) []etcdPodInfo {
 	return etcdPodInfos
 }
 
+//getMainMaster iterates over all control nodes in cluster and returns API EP node
+//return API EP node if successful, nil otherwise
 func getMainMaster(cluster *pb.K8Scluster) *pb.Node {
 	for _, nodepool := range cluster.ClusterInfo.GetNodePools() {
 		for _, node := range nodepool.Nodes {
@@ -201,6 +200,8 @@ func getMainMaster(cluster *pb.K8Scluster) *pb.Node {
 	return nil
 }
 
+//kcDrainNode executes kubectl drain node --ignore-daemonsets --delete-local-data for a specified node
+//return nil if successful, error otherwise
 func kcDrainNode(nodeName, kubeconfig string) error {
 	log.Info().Msgf("kubectl drain %s --ignore-daemonsets --delete-local-data", nodeName)
 	cmd := fmt.Sprintf("kubectl drain %s --ignore-daemonsets --delete-local-data --kubeconfig <(echo '%s')", nodeName, kubeconfig)
@@ -212,6 +213,8 @@ func kcDrainNode(nodeName, kubeconfig string) error {
 	return nil
 }
 
+//kcDeleteNode executes kubectl delete node for a specified node
+//return nil if successful, error otherwise
 func kcDeleteNode(nodeName, kubeconfig string) error {
 	log.Info().Msgf("kubectl delete node %s" + nodeName)
 	cmd := fmt.Sprintf("kubectl delete node %s --kubeconfig <(echo '%s')", nodeName, kubeconfig)
@@ -223,6 +226,8 @@ func kcDeleteNode(nodeName, kubeconfig string) error {
 	return nil
 }
 
+//kcGetNodeNames will find a node names for a particular cluster
+//return slice of node names and nil if successful, nil and error otherwise
 func kcGetNodeNames(kubeconfig string) ([]string, error) {
 	nodesQueryCmd := fmt.Sprintf("kubectl --kubeconfig <(echo \"%s\") get nodes -n kube-system --no-headers -o custom-columns=\":metadata.name\" ", kubeconfig)
 	output, err := exec.Command("bash", "-c", nodesQueryCmd).CombinedOutput()
@@ -231,4 +236,26 @@ func kcGetNodeNames(kubeconfig string) ([]string, error) {
 		return nil, err
 	}
 	return strings.Split(string(output), "\n"), nil
+}
+
+//getNodesToDelete chooses a particular nodes which will be deleted based on the clusterNodesToDelete values
+//return slice of nodes and etcd members to delete
+func getNodesToDelete(clusterNodesToDelete *nodesToDelete, nodepools []*pb.NodePool) (nodesToDelete []string, etcdToDelete []string) {
+	for _, nodepool := range nodepools {
+		for i := len(nodepool.Nodes) - 1; i >= 0; i-- {
+			// get count from nodepool
+			count, ok := clusterNodesToDelete.nodes[nodepool.Name]
+			// if nodepool found in clusterNodesToDelete && count to delete is non zero -> pick a node to delete
+			if count.Count > 0 && ok {
+				count.Count--
+				nodesToDelete = append(nodesToDelete, nodepool.Nodes[i].GetName())
+				log.Info().Msgf("Choosing node %s, with public IP %s, private IP %s for deletion\n", nodepool.Nodes[i].GetName(), nodepool.Nodes[i].GetPublic(), nodepool.Nodes[i].GetPrivate())
+				//if nodepool is control, append it to etcdToDelete
+				if nodepool.Nodes[i].NodeType > pb.NodeType_worker {
+					etcdToDelete = append(etcdToDelete, nodepool.Nodes[i].GetName())
+				}
+			}
+		}
+	}
+	return nodesToDelete, etcdToDelete
 }
